# getmetricdata refactor QueryID generation and result mapping