--- a/pkg/job/getmetricdata/processor.go
+++ b/pkg/job/getmetricdata/processor.go
@@ -4,7 +4,8 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"sync"
+	"strconv"
+	"strings"
 
 	"golang.org/x/sync/errgroup"
 
@@ -14,7 +15,7 @@ import (
 )
 
 type Client interface {
-	GetMetricData(ctx context.Context, logger logging.Logger, getMetricData []*model.CloudwatchData, namespace string, length int64, delay int64, configuredRoundingPeriod *int64) []cloudwatch.MetricDataResult
+	GetMetricData(ctx context.Context, logger logging.Logger, data []*model.CloudwatchData, namespace string, length int64, delay int64, configuredRoundingPeriod *int64) []cloudwatch.MetricDataResult
 }
 
 type Processor struct {
@@ -35,11 +36,9 @@ func (p Processor) Run(ctx context.Context, logger logging.Logger, namespace str
 	metricDataLength := len(requests)
 	partitionSize := int(math.Ceil(float64(metricDataLength) / float64(p.metricsPerQuery)))
 	logger.Debug("GetMetricData partitions", "size", partitionSize)
-	getMetricDataOutput := make([][]cloudwatch.MetricDataResult, 0, partitionSize)
 
 	g, gCtx := errgroup.WithContext(ctx)
 	g.SetLimit(p.concurrency)
-	mu := sync.Mutex{}
 	count := 0
 	for i := 0; i < metricDataLength; i += p.metricsPerQuery {
 		start := i
@@ -51,12 +50,10 @@ func (p Processor) Run(ctx context.Context, logger logging.Logger, namespace str
 		count++
 
 		g.Go(func() error {
-			input := requests[start:end]
+			input := addQueryIDsToBatch(requests[start:end])
 			data := p.client.GetMetricData(gCtx, logger, input, namespace, jobMetricLength, jobMetricDelay, jobRoundingPeriod)
 			if data != nil {
-				mu.Lock()
-				getMetricDataOutput = append(getMetricDataOutput, data)
-				mu.Unlock()
+				mapResultsToBatch(logger, data, input)
 			} else {
 				logger.Warn("GetMetricData partition empty result", "start", start, "end", end, "partitionNum", partitionNum)
 			}
@@ -69,8 +66,6 @@ func (p Processor) Run(ctx context.Context, logger logging.Logger, namespace str
 		return nil, fmt.Errorf("GetMetricData work group error: %w", err)
 	}
 
-	mapResultsToMetricDatas(getMetricDataOutput, requests, logger)
-
 	// Remove unprocessed/unknown elements in place, if any. Since getMetricDatas
 	// is a slice of pointers, the compaction can be easily done in-place.
 	requests = compact(requests, func(m *model.CloudwatchData) bool {
@@ -80,45 +75,41 @@ func (p Processor) Run(ctx context.Context, logger logging.Logger, namespace str
 	return requests, nil
 }
 
-func mapResultsToMetricDatas(output [][]cloudwatch.MetricDataResult, datas []*model.CloudwatchData, logger logging.Logger) {
-	// queryIDToData is a support structure used to easily find via a QueryID, the corresponding
-	// model.CloudatchData.
-	queryIDToData := make(map[string]*model.CloudwatchData, len(datas))
-
-	// load the index
-	for _, data := range datas {
-		queryIDToData[data.GetMetricDataProcessingParams.QueryID] = data
+func addQueryIDsToBatch(batch []*model.CloudwatchData) []*model.CloudwatchData {
+	for i, entry := range batch {
+		entry.GetMetricDataProcessingParams.QueryID = indexToQueryID(i)
 	}
 
-	// Update getMetricDatas slice with values and timestamps from API response.
-	// We iterate through the response MetricDataResults and match the result ID
-	// with what was sent in the API request.
-	// In the event that the API response contains any ID we don't know about
-	// (shouldn't really happen) we log a warning and move on. On the other hand,
-	// in case the API response does not contain results for all the IDs we've
-	// requested, unprocessed elements will be removed later on.
-	for _, data := range output {
-		if data == nil {
+	return batch
+}
+
+func mapResultsToBatch(logger logging.Logger, results []cloudwatch.MetricDataResult, batch []*model.CloudwatchData) {
+	for _, entry := range results {
+		id, err := queryIDToIndex(entry.ID)
+		if err != nil {
+			logger.Warn("GetMetricData returned unknown Query ID", "err", err, "query_id", id)
 			continue
 		}
-		for _, metricDataResult := range data {
-			// find into index
-			metricData, ok := queryIDToData[metricDataResult.ID]
-			if !ok {
-				logger.Warn("GetMetricData returned unknown metric ID", "metric_id", metricDataResult.ID)
-				continue
+		if batch[id].GetMetricDataResult == nil {
+			cloudwatchData := batch[id]
+			cloudwatchData.GetMetricDataResult = &model.GetMetricDataResult{
+				Statistic: cloudwatchData.GetMetricDataProcessingParams.Statistic,
+				Datapoint: entry.Datapoint,
+				Timestamp: entry.Timestamp,
 			}
-			// skip elements that have been already mapped but still exist in queryIDToData
-			if metricData.GetMetricDataResult != nil {
-				continue
-			}
-			metricData.GetMetricDataResult = &model.GetMetricDataResult{
-				Statistic: metricData.GetMetricDataProcessingParams.Statistic,
-				Datapoint: metricDataResult.Datapoint,
-				Timestamp: metricDataResult.Timestamp,
-			}
+
 			// All GetMetricData processing is done clear the params
-			metricData.GetMetricDataProcessingParams = nil
+			cloudwatchData.GetMetricDataProcessingParams = nil
 		}
 	}
+}
+
+func indexToQueryID(i int) string {
+	return fmt.Sprintf("id_%d", i)
+}
+
+func queryIDToIndex(queryID string) (int, error) {
+	noID := strings.TrimPrefix(queryID, "id_")
+	id, err := strconv.Atoi(noID)
+	return id, err
 }